BUG: test, fix flexible dtype conversion on class with __array__
